Fix for endles build loop on npm run dev
If you do ‘npm run dev’ when there is a interaction in src/interactions you get in to a endles loop of constantly rebuilding the set and it never gets to serving the set. If you do ‘npm rund build’ and then ‘npm run serve’ stil works.
The problem seems to be that when you build the set a javascript version of the interactions gets stored in src/interactions. And because nodemon is watching the src directory when you do ‘npm run dev’ is sees the creation of the javascript file as a file change and restarts the build process and that creates the file again. So it’s a endless loop.